requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-amazon==8.14.0
apache-airflow-providers-apache-spark==4.11.0
boto3==1.34.122
pyspark==3.5.1
python-dotenv==1.0.1